lazy synchronize signatures
Previously, the sid-msg.map was synchronized with the database on
startup. This involved selecting every signature from the database (look
this up: did it limit to one per (sid,gid)???), adding it to a cache,
then adding everything in the sid-msg.map not present in the cache to
the database.

When two sensors were started concurrently, duplicate entries were
inserted.

This change makes signature synchronization happen on demand when alerts
are inserted into the database. When an alert is inserted, the database
cache is checked for any cached signatures. If there is a match, the
event is updated with the db_id from the cache...

If there is not a match, we check the database to see if that signature
has been inserted in the database. If it has, we pull that and update
the cache.

If there was not a match, we check the sid-msg.map for a match. If the
sid-msg.map contains a match, then the signature is inserted into the
database and database cache.

If nothing matches, the signature is inserted with the default name of
'Snort Alert [gid:sid:rev]'

This commit completely breaks reference caching, which was also
previously done on startup. This will need to be moved to a function
that Synchronizes a single signature.