replace cacheSignatureObj with a hash table (khash)
cacheSignatureObj previously formed a linked list of dbSignatureObjs.
This replaces the linked list with a hash table using the khash library.
Now cacheSignatureHead is a khash of khashes. cacheSignatureHead uses
gid as the key while the hash tables within cacheSignatureHead use sid
as the key. [This basically forms a multidimensional hash table:
signature = cacheSignatureHead[gid][sid]).